update sequence for all objects

+++ src/main/java/org/olegmell/repository/ServicesRepository.java
@@ -3,6 +3,6 @@ package org.olegmell.repository;
 import org.olegmell.domain.Services;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ServicesRepository extends CrudRepository<Services, Long> {
+public interface ServicesRepository extends CrudRepository<Services, Integer> {
 
 }
